refactor the code according to k8s standards
