+++ config/crd/bases/pinot.apache.io_pinots.yaml
@@ -6,9 +6,9 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.4.0
   creationTimestamp: null
-  name: pinots.operators.apache.io
+  name: pinots.pinot.apache.io
 spec:
-  group: operators.apache.io
+  group: pinot.apache.io
   names:
     kind: Pinot
     listKind: PinotList
