+++ config/crd/bases/pinot.apache.io_tables.yaml
@@ -6,9 +6,9 @@ metadata:
   annotations:
     controller-gen.kubebuilder.io/version: v0.4.0
   creationTimestamp: null
-  name: tables.operators.apache.io
+  name: tables.pinot.apache.io
 spec:
-  group: operators.apache.io
+  group: pinot.apache.io
   names:
     kind: Table
     listKind: TableList
