+++ config/crd/patches/cainjection_in_schemas.yaml
@@ -5,4 +5,4 @@ kind: CustomResourceDefinition
 metadata:
   annotations:
     cert-manager.io/inject-ca-from: $(CERTIFICATE_NAMESPACE)/$(CERTIFICATE_NAME)
-  name: schemas.operators.apache.io
+  name: schemas.pinot.apache.io
