+++ config/crd/patches/cainjection_in_tables.yaml
@@ -5,4 +5,4 @@ kind: CustomResourceDefinition
 metadata:
   annotations:
     cert-manager.io/inject-ca-from: $(CERTIFICATE_NAMESPACE)/$(CERTIFICATE_NAME)
-  name: tables.operators.apache.io
+  name: tables.pinot.apache.io
