+++ config/crd/patches/cainjection_in_tenant.yaml
@@ -5,4 +5,4 @@ kind: CustomResourceDefinition
 metadata:
   annotations:
     cert-manager.io/inject-ca-from: $(CERTIFICATE_NAMESPACE)/$(CERTIFICATE_NAME)
-  name: tenants.operators.apache.io
+  name: tenants.pinot.apache.io
