+++ config/crd/patches/webhook_in_schemas.yaml
@@ -3,7 +3,7 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: schemas.operators.apache.io
+  name: schemas.pinot.apache.io
 spec:
   conversion:
     strategy: Webhook
