+++ config/crd/patches/webhook_in_tables.yaml
@@ -3,7 +3,7 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: tables.operators.apache.io
+  name: tables.pinot.apache.io
 spec:
   conversion:
     strategy: Webhook
