+++ config/crd/patches/webhook_in_tenant.yaml
@@ -3,7 +3,7 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: tenants.operators.apache.io
+  name: tenants.pinot.apache.io
 spec:
   conversion:
     strategy: Webhook
