+++ config/rbac/pinot_editor_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: pinot-editor-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots
   verbs:
@@ -17,7 +17,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots/status
   verbs:
