+++ config/rbac/pinot_viewer_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: pinot-viewer-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots
   verbs:
@@ -13,7 +13,7 @@ rules:
   - list
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots/status
   verbs:
