+++ config/rbac/role.yaml
@@ -45,7 +45,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots
   verbs:
@@ -55,13 +55,13 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots/finalizers
   verbs:
   - update
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - pinots/status
   verbs:
@@ -69,7 +69,7 @@ rules:
   - patch
   - update
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - schemas
   verbs:
@@ -81,7 +81,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - schemas/status
   verbs:
@@ -89,7 +89,7 @@ rules:
   - patch
   - update
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tables
   verbs:
@@ -101,7 +101,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tables/status
   verbs:
@@ -109,7 +109,7 @@ rules:
   - patch
   - update
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tenants
   - tenants/finalizers
@@ -122,7 +122,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tenants/status
   verbs:
