+++ config/rbac/schema_editor_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: schema-editor-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - schemas
   verbs:
@@ -17,7 +17,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - schemas/status
   verbs:
