+++ config/rbac/schema_viewer_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: schema-viewer-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - schemas
   verbs:
@@ -13,7 +13,7 @@ rules:
   - list
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - schemas/status
   verbs:
