+++ config/rbac/table_editor_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: table-editor-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tables
   verbs:
@@ -17,7 +17,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tables/status
   verbs:
