+++ config/rbac/table_viewer_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: table-viewer-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tables
   verbs:
@@ -13,7 +13,7 @@ rules:
   - list
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tables/status
   verbs:
