+++ config/rbac/tenant_editor_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: tenant-editor-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tenants
   verbs:
@@ -17,7 +17,7 @@ rules:
   - update
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tenants/status
   verbs:
