+++ config/rbac/tenant_viewer_role.yaml
@@ -5,7 +5,7 @@ metadata:
   name: tenant-viewer-role
 rules:
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tenants
   verbs:
@@ -13,7 +13,7 @@ rules:
   - list
   - watch
 - apiGroups:
-  - operators.apache.io
+  - pinot.apache.io
   resources:
   - tenants/status
   verbs:
