+++ controllers/pinot/watches.go
@@ -3,7 +3,7 @@ package pinot
 import (
 	"errors"
 
-	pinotv1alpha1 "github.com/spaghettifunk/pinot-operator/api/pinot/v1alpha1"
+	pinotv1alpha1 "github.com/spaghettifunk/pinot-operator/pkg/apis/pinot/v1alpha1"
 	"github.com/spaghettifunk/pinot-operator/pkg/crds"
 	"github.com/spaghettifunk/pinot-operator/pkg/k8sutil"
 	appsv1 "k8s.io/api/apps/v1"
