+++ docs/examples/hello-world/pinot.yaml
@@ -1,4 +1,4 @@
-apiVersion: operators.apache.io/v1alpha1
+apiVersion: pinot.apache.io/v1alpha1
 kind: Pinot
 metadata:
   name: pinot-sample
