+++ docs/examples/tenants/server_tenant.yaml
@@ -1,4 +1,4 @@
-apiVersion: operators.apache.io/v1alpha1
+apiVersion: pinot.apache.io/v1alpha1
 kind: Tenant
 metadata:
   name: server-tenant-sample
