+++ pkg/apis/addtoscheme_pinot_v1alpha1.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	"github.com/spaghettifunk/pinot-operator/api/pinot/v1alpha1"
+	"github.com/spaghettifunk/pinot-operator/pkg/apis/pinot/v1alpha1"
 )
 
 func init() {
