+++ pkg/apis/pinot/v1alpha1/groupversion_info.go
@@ -14,12 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package v1alpha1 contains API Schema definitions for the operators v1alpha1 API group
+// Package v1alpha1 contains API Schema definitions for the pinot v1alpha1 API group
 // +kubebuilder:object:generate=true
-// +groupName=operators.apache.io
+// +groupName=pinot.apache.io
 // +k8s:openapi-gen=true
 // +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=github.com/spaghettifunk/pinot-operator/api
+// +k8s:conversion-gen=github.com/spaghettifunk/pinot-operator/pkg/apis
 // +k8s:defaulter-gen=TypeMeta
 package v1alpha1
 
@@ -30,7 +30,7 @@ import (
 
 var (
 	// GroupVersion is group version used to register these objects
-	GroupVersion = schema.GroupVersion{Group: "operators.apache.io", Version: "v1alpha1"}
+	GroupVersion = schema.GroupVersion{Group: "pinot.apache.io", Version: "v1alpha1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: GroupVersion}
