+++ pkg/apis/pinot/v1alpha1/schema_types.go
