+++ pkg/resources/controller/controller.go
@@ -1,7 +1,7 @@
 package controller
 
 import (
-	pinotv1alpha1 "github.com/spaghettifunk/pinot-operator/api/pinot/v1alpha1"
+	pinotv1alpha1 "github.com/spaghettifunk/pinot-operator/pkg/apis/pinot/v1alpha1"
 
 	"github.com/go-logr/logr"
 	"github.com/goph/emperror"
